Ridge and Lasso Regression
A US-based housing company named Surprise Housing has decided to enter the Australian market. The company uses data analytics to purchase houses at a price below their actual value and flip them at a higher price. For the same purpose, the company has collected a data set from house sales in Australia. The data is provided in the csv file below.
The company is looking at prospective properties to buy to enter the market.
You are required to build a regression model using regularization, so as to predict the actual value of the prospective properties and decide whether to invest in them or not.
The company wants to know:
1. Which variables are significant in predicting the price of a house
2. How well those variables describe the price of a house