Support Vector Machine
A classic problem in the field of pattern recognition is that of handwritten digit recognition. Suppose that you have images of handwritten digits ranging from 0-9 written by various people in boxes of a specific size - similar to the application forms in banks and universities.
The goal is to develop a model that can correctly identify the digit (between 0-9) written in an image. 

Data Links:
https://www.kaggle.com/c/digit-recognizer/data